Extract ProjectCreationPanel from Main
